Move everything into single script
I had problems getting this working with Drone, and it took me a while
to dawn on me that Drone will _not_ respect the WORKDIR because it
specifically sets this to be able to use the shared mounted volume.
Instead, Drone plugins should be expected to just have binaries or
scripts available in the `PATH` to perform specific functions.

For this reason, I have moved everything into a single script, and just
copy this script over in the container, specifically installing the only
dependency we require (httparty).

I am keeping the Gemfile for local development only.