Show first line of commit message in notification
This adds the commit title to the build notification; for nosey types
like me this is often quite interesting to see in the notifications
channel, giving a bit more context to changes being made than the branch
name.

To make room for it I've moved the build time into the status panel.