Ensure LANG environment variable is set
There was an instance where the Slack notification failed for an
otherwise successful build:

```
/bin/drone-slack:38:in `encode': "\xE2" from ASCII-8BIT to UTF-8 (Encoding::UndefinedConversionError)
    from /bin/drone-slack:38:in `to_json'
    from /bin/drone-slack:38:in `template'
    from /bin/drone-slack:44:in `notify'
    from /bin/drone-slack:165:in `<main>'
```

It seems because the committer was using fancy characters in their
commit message and the locale was not set, so Ruby blew up.

This ensures the LANG environment variable is set to protect against
people using fancy characters.